+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ v0.6 (XXXX/XX/XX)
 -----------------
 Philip Loche, Simon Gravelle, Srihas Velpuri, Henrik Jaeger
 
+- Use base units exclusively (!115)
 - Higher tolerance for non-neutral systems (1E-4 instead of 1E-5)
 - `charge`neutral decorator uses `check_compound` now
 - Add option to symmetrize profiles using `ProfilePlanarBase` (!116)